feature(scylla operator): add events for catching pods wrong scheduling
We have fixed a lot of bugs related to the wrong scheduling of non-scylla pods on K8S.
Main problem: we can get some non-scylla pods be scheduled onto scylla nodes and it may
lead to unpredicted errors when we run different nemesis'.

So, we should not only fix but also keep an eye on the non-scylla pods that must not be placed
on scylla nodes.

As a result, need to create event(s) that will periodically check pods on scylla nodes and verify
that there are only allowed ones.

This commit introduce an error event.

Task: https://trello.com/c/cqx1W4c6/3359-k8s-add-events-for-catching-wrong-
scheduling-of-non-scylla-pods