improvement(nemesis): add wrapper to disrupt method during init
This commit is part of support operator multi-tenant longevity.
Longevity test should be run on every tenant in parallel threads.
In the current realization a disrupt wrapper is applied on the class level. As result the wrapper
will be applied a few times (as much as clusters amount)

Move applying disrupt wrapper inside __init__ function to prevent multipying.