test(longevity): add longevity with rf=1
This PR is trying to cover customer issue https://github.com/scylladb/scylladb/issues/16759.
This is from the issue description:

After Upgrading Scylla from 5.2.11 via 5.4.0 to 5.4.1, we started observing missing data in Scylla.
Every once in a while, an INSERT is committed successfully but the inserted values are not visible
until Scylla is restarted.
As far as we can tell, version 5.2.11 was not affected while 5.4.1 is. 5.4.0 was not long enough
in operation to make a reliable statement.

The suspected bug occurs extremely rarely, making it hard for us to reproduce.
Out of ~800M inserts per day, only a dozen are affected.

The load is run on keyspace with RF 1 SCT test.
We use scylla-bench because its validate data better then in cassandra-stress